# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NORA API
  version: 0.2.0
  description: |
    API REST do backend NORA (Spring Boot 3 / Java 21). Cobre autenticacao com cookies httpOnly,
    IAM estilo AWS (grupos + policies + attachments + audit), convites por e-mail, contexto comercial
    do tenant, dominio corporativo, upload/listagem/detalhe de reunioes, gestao de tarefas extraidas,
    analise ao vivo e emissao de token efemero para Azure Speech.

    ## Autenticacao
    A maioria dos endpoints exige um JWT. Ha dois modos de transporte equivalentes:

    - **Bearer** — header `Authorization: Bearer <accessToken>` (esquema `bearerAuth`).
    - **Cookie httpOnly** — `POST /auth/login` (e `POST /iam/invites/{token}/accept`) define
      `Set-Cookie: nora_access` (`Path=/`, `SameSite=Lax`, `HttpOnly`) e `Set-Cookie: nora_refresh`
      (`Path=/auth`, `SameSite=Strict`, `HttpOnly`). O navegador reenvia `nora_access` automaticamente.
      `nora_refresh` so chega nas rotas sob `/auth` (refresh/logout). A flag `Secure` e ligada em prod
      via `nora.auth.cookie-secure`.

    O campo `accessToken` continua presente no corpo JSON do login por compat com clientes legados;
    novos clientes devem confiar nos cookies.

    ## Autorizacao IAM (ADR 0007)
    NAO existem papeis fixos. A autorizacao e baseada em **policies IAM** (estilo AWS) avaliadas por
    `(action, resource)`. Cada endpoint protegido declara na descricao a action IAM exigida
...
    `403 IAM_FORBIDDEN`.

    ## Formato de erro
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      security: []
      summary: Cria conta (e tenant) e dispara verificacao de e-mail
      description: |
        Publico. Rate-limited por IP (`429 RATE_LIMITED`, header `Retry-After`). Cria o usuario e o
        tenant associado. Em ambientes nao-produtivos o token de verificacao volta em
        `emailVerificationDevToken` para facilitar testes.

        O endpoint responde `201` com o mesmo formato de corpo para qualquer e-mail, inclusive um
        que ja tenha conta: por ser publico e nao autenticado, ele nao confirma quem esta
        cadastrado. Nesse caso nada e criado e o dono do endereco recebe um aviso por e-mail. O
        `409 EMAIL_ALREADY_TAKEN` permanece apenas no aceite de convite (`/iam/invites/{id}/accept`),
        onde quem chama ja tem credencial do workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Requisicao aceita (verificacao de e-mail pendente)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignupResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/verify-email:
    post:
      tags: [Auth]
      security: []
      summary: Confirma o e-mail a partir do token de verificacao
      description: Publico. Token invalido/usado/expirado retorna `400 TOKEN_INVALID`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyEmailRequest' }
      responses:
        '204': { description: E-mail verificado }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login com e-mail e senha
      description: |
        Publico. Rate-limited por IP. Em caso de sucesso, alem do corpo `LoginResponse`, responde com
        dois headers `Set-Cookie`:

        - `nora_access` — JWT de acesso curto. `Path=/`, `SameSite=Lax`, `HttpOnly`.
        - `nora_refresh` — refresh opaco de longa duracao. `Path=/auth`, `SameSite=Strict`, `HttpOnly`.

        Credenciais invalidas retornam `401 INVALID_CREDENTIALS`; e-mail nao verificado
        `401 EMAIL_NOT_VERIFIED`; conta desativada `403 USER_DISABLED`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Autenticado. Define cookies `nora_access` e `nora_refresh`.
          headers:
            Set-Cookie:
              description: Cookies httpOnly `nora_access` (Path=/) e `nora_refresh` (Path=/auth).
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Renova o access token (rotaciona o refresh)
      description: |
        Le o refresh do cookie `nora_refresh` (Path=/auth); se ausente, aceita `Authorization: Bearer
        <refresh>`. Faz rotacao: revoga o refresh anterior e reescreve ambos os cookies
        (`nora_access` + `nora_refresh`). O novo access token sai apenas no cookie; o corpo traz
        somente metadados. Ausente/expirado/revogado retorna `401 REFRESH_TOKEN_INVALID`.
      responses:
        '200':
          description: Tokens renovados. Reescreve cookies `nora_access` e `nora_refresh`.
          headers:
            Set-Cookie:
              description: Novos cookies httpOnly `nora_access` e `nora_refresh`.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefreshResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/password/reset/request:
    post:
      tags: [Auth]
      security: []
      summary: Solicita reset de senha
      description: |
        Publico. Rate-limited por e-mail. Responde sempre `202` com mensagem indistinguivel para nao
        vazar quais e-mails existem. Em ambientes nao-produtivos o token volta em
        `passwordResetDevToken`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestPasswordResetRequest' }
      responses:
        '202':
          description: Aceito (resposta indistinguivel exista ou nao a conta)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestPasswordResetResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/password/reset/confirm:
    post:
      tags: [Auth]
      security: []
      summary: Confirma o reset de senha com o token
      description: Publico. Token invalido/usado/expirado retorna `400 TOKEN_INVALID`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmPasswordResetRequest' }
      responses:
        '204': { description: Senha redefinida }
        '400': { $ref: '#/components/responses/BadRequest' }

  # ----------------------------------------------------------------------------
  # IAM — groups
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: nora_access
      description: |
        Cookie httpOnly definido por `POST /auth/login` (e por `POST /iam/invites/{token}/accept`).
        Equivalente ao Bearer; o navegador o reenvia automaticamente. Renovado via `POST /auth/refresh`.
...
  responses:
    BadRequest:
      description: Requisicao invalida (`VALIDATION_FAILED`, `MALFORMED_REQUEST`, `TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Nao autenticado (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`, `REFRESH_TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Sem permissao (`FORBIDDEN`, `IAM_FORBIDDEN`, `USER_DISABLED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Rate limit excedido (`RATE_LIMITED`, `RATE_LIMIT_EXCEEDED`). Header `Retry-After`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code: { type: string, example: MEETING_NOT_FOUND }
        message: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }

    # ---- shared enums ----
...
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 128 }
        displayName: { type: string, maxLength: 120 }
    SignupResponse:
      type: object
      required: [userId, tenantId, message]
      properties:
        userId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        message: { type: string }
        emailVerificationDevToken:
          type: string
          description: Presente apenas em ambientes nao-produtivos.
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, maxLength: 128 }
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresInSeconds, userId, tenantId, email, displayName]
      description: |
        Corpo do login (e do aceite de convite). Tambem acompanha headers `Set-Cookie`
        (`nora_access`, `nora_refresh`). `accessToken`/`refreshToken` no corpo existem por compat
        legada; novos clientes devem confiar nos cookies httpOnly.
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer, format: int64, description: TTL do access token em segundos. }
        userId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
    RefreshResponse:
      type: object
      required: [tokenType, expiresInSeconds]
      description: |
        O novo access token sai no cookie `nora_access`; o corpo traz apenas metadados para o cliente
        agendar o proximo refresh.
      properties:
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer, format: int64 }
    RequestPasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    RequestPasswordResetResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        passwordResetDevToken:
          type: string
          description: Presente apenas em ambientes nao-produtivos.
    ConfirmPasswordResetRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string }
        newPassword: { type: string, minLength: 10, maxLength: 128 }

    # ---- IAM ----
...